+++ docs/whatsnew/changelog.rst
@@ -20,6 +20,7 @@ New features
 ~~~~~~~~~~~~
 .. Add here new public features (do not delete this comment)
 
+* Add Stejskal-Tanner kernel for 2D IRIS
 * Fancy indexing using location now supported.
 * Add an example for NMR processing of a series of CP-MAS spectra.
 * Add an example for processing NMR relaxation data